test: separate test by datasource
